wifi: ieee80211: split mesh definitions out
[ Upstream commit 69674282fc97fffd98a85ab5b4837edbc5898145 ]

The ieee80211.h file has gotten very long, start splitting it
by putting mesh definitions into a separate file.

Link: https://patch.msgid.link/20251105153843.489713ca8b34.I3befb4bf6ace0315758a1794224ddd18c4652e32@changeid
Stable-dep-of: cb0caadb64ca ("wifi: ieee80211: fix definition of EHT-MCS 15 in MRU")
(cherry picked from commit 3f459076b2d823cd8e52b701f7259236bde4d4bc)
